Added sensor_subscribe + delay_async, both have been preliminarily tested, but a more thorough vetting is ongoing.
Notes:

g_track and g_sensor are now the canonical source for track
information.  The former is an array of nodes and the latter is an array
of pointers to nodes which are sensors indexed by sensor number. If
either of these is required simply include track.h. Currently the
variables get initialized in the initTask to track a but they can be
re-initialized by the console. When possible prefer passing around
track_node pointers rather than array indices this should usually be
possible and this is what the routing and sensor APIs provide anyway.

sensor_subscribe() works as discussed. The sensor printer was
re-factored to use the new API and has been moved into sensor.c which
gets initialized by init_sensors() in initTask.

kern_print was added to the kernel and is now used by dbgt, this allows
for uninterrupted synchronous debugging output and also prepends the
currently printing task number for convenience.

delay_until_async and delay_async work similarly to sensor_subscribe,
they return a td which can be compared against the one returned by the
receive call to differentiate between different requests.

The train stuff hasn't been touched or tested and the console commands
may or may not be broken :/.